refactor(artwork): move page to construction message
BREAKING CHANGE: Temporarily removes detailed artwork display functionality
